+++ lib/concolor/extension.rb
@@ -1,7 +1,17 @@
 # frozen_string_literal: true
 
 module Concolor
-  # Extension module for String class
+  # Extension module
+  #
+  # @example
+  #   class CustomString < String
+  #     include Concolor::Extension
+  #   end
+  #
+  #   CustomString.new('Some text').concolor(:bold, :red)
+  #   # => "\033[31;1mSome text\033[0m"
+  #
+  # @since 0.0.1
   module Extension
     # @param options [Array<Symbol>]
     # @return [Concolor::String]