+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # `Concolor` Changelog
 
-## [0.0.1] - 2015-05-22
+## Development
+
+### Added
+
+- `nil`
+
+### Changed
+
+- deprecated `Concolor::Version.current`
+- removed `Concolor::Mappers::*` modules and added new constants:
+  - `Concolor::Mappers::COLORS`
+  - `Concolor::Mappers::BG_COLORS`
+  - `Concolor::Mappers::MODES`
+
+### Fixed
+
+- `nil`
+
+---
+
+## [1.0.0] - October 23, 2022
+
+### Added in `1.0.0`
+
+- new `Concolor::Mappers::*` for bg-color, color, mode
+- new class `Concolor::String`
+- new module `Concolor::Extension`
+- `.devcontainer` feature for `VS Code`
+- `demo` folder with example
+
+### Changed in `1.0.0`
+
+- `nil`
+
+### Fixed in `1.0.0`
+
+- `nil`
+
+---
+
+## [0.0.1] - May 22, 2015
 
 - Initial release